+++ Changelog.md
@@ -1,5 +1,9 @@
 -------------------------
-CHANGELOG - 9/30/2024 | PUSHED
+CHANGELOG - 10/1/2024 | PUSHED 10/1/2024
+* Fixed improper distance formula (Swapped to built-in function)
+* Changed Proxy priority. It will now happen before any other items.
+-------------------------
+CHANGELOG - 9/30/2024 | PUSHED 9/30/2024
 * Fixed improper queue calls (never called onChatRecieve,
     improper message being edited)
 * Fixed Proxies not disabling each other when activating multiple